requirements.txt
gradio
ultralytics
numpy
pillow
opencv-python-headless




